requirements.txt
fastapi>=0.128.3,<0.129.0
uvicorn>=0.30.0,<1.0.0
debugpy>=1.8.0,<2.0.0
google-cloud-storage
python-multipart
haversine
pillow
ultralytics

--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision